Fixed the inclusion of Eigen3, fixed Assertions and replaced M_PI with std::numbers::pi

+++ src/embeddingLib/include/dVec/VecList.hpp
@@ -40,7 +40,7 @@ class VecList {
 
     void setAll(double default_value) {
         #pragma omp parallel for
-        for (size_t i = 0; i < size(); ++i) {
+        for (int64_t i = 0; i < size(); ++i) {
             (*this)[i].setAll(default_value);
         }
     }
